renamed intergation_tests to avoid import problems on some systems
