+++ .travis.yml
@@ -34,7 +34,7 @@ install:
 
 script:
   - py.test unittests
-  - py.test integration_tests
+  - py.test fec_integration_tests
   - flake8 spinn_front_end_common
   - flake8 unittests integration_tests
   - find spinn_front_end_common -name '*.xml' | xargs -n 1 support/validate-xml.sh
