+++ fec_integration_tests/interface/interface_functions/test_front_end_common_load_executable_images.py
