+++ fec_integration_tests/mock_machine.py
